Update the Cropped Image Renderer
Hey, thanks so much for your diligence and attention in trying to help me find a solution for a higher quality render of the cropped image. I know that's not really under the scope of the utility and really appreciated the effort. I did a bit of tinkering and came up with this, which is a pixel-perfect solution. I'm no canvas expert so I can't really explain why it works better than the previous version... but I did spend a few hours trying various solutions and this was the one that worked best.